first

+++ Game/src/Player.h
@@ -3,13 +3,13 @@
 // Copyright(c) 2020 Arthur Bacon and Kevin Dill
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this softwareand associated documentation files(the "Software"), to deal
+// of this software and associated documentation files(the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and /or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions :
 // 
-// The above copyright noticeand this permission notice shall be included in all
+// The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
 // 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
@@ -48,10 +48,10 @@ public:
     const std::vector<Entity*>& getBuildings() const { return m_Buildings; }
     const std::vector<Entity*>& getMobs() const { return m_Mobs; }
 
-    virtual unsigned int getNumBuildings() const { return (unsigned int)m_Buildings.size(); }
+    virtual unsigned int getNumBuildings() const { return m_Buildings.size(); }
     virtual EntityData getBuilding(unsigned int i) const;
 
-    virtual unsigned int getNumMobs() const { return (unsigned int)m_Mobs.size(); }
+    virtual unsigned int getNumMobs() const { return m_Mobs.size(); }
     virtual EntityData getMob(unsigned int i) const;
 
     virtual unsigned int getNumOpponentBuildings() const { return GetOpponent().getNumBuildings(); }
